pyproject.toml
[tool.poetry]
name = "nerf-gun-control"
version = "0.1.0"
description = "A Twitch bot for controlling a Nerf gun system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "nerf_gun_control"}]

[tool.poetry.dependencies]
python = "^3.11"
twitchio = "^2.6.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "nerf_gun_control.main:main"

[tool.black]
line-length = 100
target-version = ['py310']
extend-exclude = '''
/(
  # directories
  \.conda
)/
'''

[tool.flake8]
max-line-length = 100
exclude = [ ".venv", ".git", "__pycache__", ".conda/" ]
ignore = "E501"

